Cleanup logic and remove deprecated/unused attachment endpoints
