--- api/src/services/applications/delete_application_attachment.py
@@ -1,43 +1,0 @@
-import logging
-import uuid
-
-import grants_shared.adapters.db as db
-from grants_shared.util import file_util
-
-from src.auth.endpoint_access_util import check_user_access
-from src.constants.lookup_constants import Privilege
-from src.db.models.user_models import User
-from src.services.applications.get_application_attachment import get_application_attachment
-
-logger = logging.getLogger(__name__)
-
-
-def delete_application_attachment(
-    db_session: db.Session,
-    application_id: uuid.UUID,
-    application_attachment_id: uuid.UUID,
-    user: User,
-) -> None:
-    """Delete an application attachment"""
-    # Fetch the application attachment, handles verifying
-    # it exists, and that the user has access attachment.
-    application_attachment = get_application_attachment(
-        db_session, application_id, application_attachment_id, user
-    )
-    # Check privileges
-    check_user_access(
-        db_session, user, {Privilege.MODIFY_APPLICATION}, application_attachment.application
-    )
-    # Delete the file from s3
-    logger.info("Deleting application attachment from s3")
-    file_util.delete_file(application_attachment.file_location)
-
-    # Mark the attachment as deleted
-    # We keep the record in the DB for auditing purposes
-    # But still remove the actual file
-    application_attachment.is_deleted = True
-    # Remove the s3 path to make clear it's gone.
-    application_attachment.file_location = "DELETED"
-
-    # The ATTACHMENT_DELETED audit event is intentionally not emitted here. It is
-    # emitted on form save when the attachment is diffed out of the form response.
